% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{add_creds_to_env}
\alias{add_creds_to_env}
\title{Saves the user's token in the environment}
\usage{
add_creds_to_env(creds = NULL)
}
\arguments{
\item{creds}{a list generated by a \code{get_creds_from_} function}
}
\description{
The user's token is saved in an environment inside the package for use by the package.
Package forgets the token at the end of the session.
}
\seealso{
Other token handling functions: \code{\link{add_creds_to_cache}},
  \code{\link{get_creds_from_cache}},
  \code{\link{get_creds_from_env}},
  \code{\link{get_creds_from_server}}
}
