% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{airport}
\alias{airport}
\title{Airport Reference Content.}
\usage{
airport(abbr = "YVR")
}
\arguments{
\item{abbr}{The airport abbreviation. Default is 'YVR', the Vancouver Airport}
}
\value{
httr content object
}
\description{
Airport Reference Content.
}
