% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{Authorization}
\alias{Authorization}
\title{Format of the authorization part for obtaining content}
\usage{
Authorization()
}
\value{
authorization string to be sent with a GET requeset
}
\description{
Format of the authorization part for obtaining content
}
