% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{cache_option}
\alias{cache_option}
\title{Interprets the cache option}
\usage{
cache_option()
}
\description{
To control the way token is stored between sessions, the package
uses looks for the option \code{lufthansar_token_cache} which can be set using
\code{options(lufthansar_token_cache=...)}. The options can include
\describe{
\item{\code{TRUE}}{to store the cache in the file \code{.lufthansa-token} in the
current working directory.}
\item{\code{FALSE}}{to suppress storing the token. Without a cached file, the
token will not be remembered between R sessions.}
\item{\code{character}}{to indicate the file where the cache will be stored. This
is used to both look for the file and store it in.}
}
This function finds the option and interprets it for use by other functions.
}
