% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{caching}
\alias{caching}
\alias{interpret_cache}
\alias{get_creds_from_cache}
\alias{add_creds_to_cache}
\title{Caching the token}
\usage{
interpret_cache(cache)

get_creds_from_cache(cache = getOption("lufthansar_token_cache"))

add_creds_to_cache(creds = NULL,
  cache = getOption("lufthansar_token_cache"))
}
\arguments{
\item{cache}{A list a \code{token} and \code{expiry} component. The \code{token} is
a string provided by Lufthansa for using the API. The \code{expiry} is a \code{}}
}
\description{
Stores and retrives tokens from a cached file.
}
\section{Functions}{
\itemize{
\item \code{interpret_cache}: Interpreting the cache setting

\item \code{get_creds_from_cache}: Retrieves previously stored token information from the cache file

\item \code{add_creds_to_cache}: Saves token information to a cache file for later retrieval
}}

