% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{get_creds_from_cache}
\alias{get_creds_from_cache}
\title{Retrieves token information from cache file}
\usage{
get_creds_from_cache()
}
\description{
Retrieves token information from cache file
}
\seealso{
Other token handling functions: \code{\link{add_creds_to_cache}},
  \code{\link{add_creds_to_env}},
  \code{\link{get_creds_from_env}},
  \code{\link{get_creds_from_server}}
}
