% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{get_creds_from_env}
\alias{get_creds_from_env}
\alias{add_creds_to_env}
\title{Token in the package environment}
\usage{
get_creds_from_env()

add_creds_to_env(creds = NULL)
}
\arguments{
\item{token}{The response token}
}
\value{
The token as a string
}
\description{
Token in the package environment
}
\section{Functions}{
\itemize{
\item \code{add_creds_to_env}: Saves the user's token in the environment

The user's token is saved in an environment inside the package for use by the package.
Package forgets the token at the end of the session.
}}

\seealso{
Other token handling functions: \code{\link{get_creds_from_server}},
  \code{\link{get_token}}

Other token handling functions: \code{\link{get_creds_from_server}},
  \code{\link{get_token}}
}
