% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_status.R
\name{get_flight_status}
\alias{get_flight_status}
\title{GEt the status of a flight.}
\usage{
get_flight_status(flight_num = "AC123", dep_date = Sys.Date(),
  verbose = TRUE)
}
\arguments{
\item{flight_num}{(the flight number)}

\item{dep_date}{Date of departure.  CAn be from yesterday to 5 days in the future, default today}

\item{verbose}{prints flight data as well as outputting the httr object}
}
\value{
httr content object
}
\description{
GEt the status of a flight.
}
