% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flight_status_departure.R
\name{get_flight_status_departure}
\alias{get_flight_status_departure}
\title{GET Flight Status at Departure Airport}
\usage{
get_flight_status_departure(airport = "FRA", fromDateTime = NULL)
}
\arguments{
\item{airport}{3-letter IATA airport code (e.g. 'YVR'). Default 'FRA'.}

\item{fromDateTime}{ISO 8601 data time string. Format is YYYY-MM-DDTHH:mm. Default is the Sys.time().}
}
\value{
httr content object
}
\description{
GET Flight Status at Departure Airport
}
\examples{

get_flight_status_departure(airport = "YVR", fromDateTime = Sys.time())

}
