% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{new_token}
\alias{new_token}
\title{Generate a new token from Lufthansa.}
\usage{
new_token(client_id, client_secret)
}
\arguments{
\item{client_id}{The client ID used to register for using the API.}

\item{client_secret}{The secret provided during registration.}
}
\value{
the token as a string
}
\description{
Generate a new token from Lufthansa.
}
\seealso{
Other token handling functions: \code{\link{get_token}},
  \code{\link{set_token}}
}
