% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{set_token}
\alias{set_token}
\title{Sets the user's token.}
\usage{
set_token(token = NULL)
}
\arguments{
\item{token}{The token as a string}
}
\description{
The user's token is saved in an environment inside the package for use by the package.
Package forgets the token at the end of the session.
}
\seealso{
Other token handling functions: \code{\link{get_token}},
  \code{\link{new_token}}
}
