% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{get_creds_from_env}
\alias{get_creds_from_env}
\title{Get token information from environment}
\usage{
get_creds_from_env()
}
\value{
a named list with components \code{token} and \code{expiry}
}
\description{
This function serves as a helper function to retrieve
token information from the package environment.
}
\seealso{
Other token handling functions: \code{\link{add_creds_to_cache}},
  \code{\link{add_creds_to_env}},
  \code{\link{get_creds_from_cache}},
  \code{\link{get_creds_from_server}}
}
