% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{get_creds_from_server}
\alias{get_creds_from_server}
\title{Generate a new token from Lufthansa.}
\usage{
get_creds_from_server(client_id = Sys.getenv("LUFTHANSA_API_CLIENT_ID"),
  client_secret = Sys.getenv("LUFTHANSA_API_CLIENT_SECRET"))
}
\arguments{
\item{client_id}{The client ID used to register for using the API.}

\item{client_secret}{The secret provided during registration.}
}
\value{
the token as a string
}
\description{
Generate a new token from Lufthansa.
}
\seealso{
Other token handling functions: \code{\link{get_creds_from_env}},
  \code{\link{get_token}}
}
