% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_cred.R
\name{get_token}
\alias{get_token}
\title{Get the current token being used by the package}
\usage{
get_token()
}
\value{
The token as a string
}
\description{
Get the current token being used by the package
}
\seealso{
Other token handling functions: \code{\link{new_token}},
  \code{\link{set_token}}
}
